+++ GRDB/FetchRequest/FetchRequest.swift
@@ -168,25 +168,7 @@ extension FetchRequest {
     /// Returns the number of rows matched by the request.
     @warn_unused_result
     public func fetchCount(db: Database) -> Int {
-        guard !query.distinct else {
-            fatalError("Counting distinct fetch requests is not implemented yet.")
-        }
-        guard query.hasSingleTableSourceAndSelectsAllColumns else {
-            fatalError("Counting fetch requests with custom selection is not implemented yet.")
-        }
-        return Int.fetchOne(db, select([_SQLExpression.Count(_SQLResultColumn.Star(nil))]))!
-    }
-}
-
-
-extension _SQLSelectQuery {
-    /// Support for FetchRequest.fetchCount()
-    private var hasSingleTableSourceAndSelectsAllColumns: Bool {
-        guard let source = source else { return false }
-        guard case .Table(name: let tableName, alias: _) = source else { return false }
-        guard selection.count == 1 else { return false }
-        guard selection[0].isStar(tableName: tableName) else { return false }
-        return true
+        return Int.fetchOne(db, FetchRequest(query.countQuery))!
     }
 }
 
@@ -312,12 +294,6 @@ extension TableMapping {
     @warn_unused_result
     public static func order(sql sql: String) -> FetchRequest<Self> {
         return all().order(sql: sql)
-    }
-    
-    /// Returns a FetchRequest sorted in reversed order.
-    @warn_unused_result
-    public static func reverse() -> FetchRequest<Self> {
-        return all().reverse()
     }
     
     /// Returns a FetchRequest which fetches *limit* rows, starting at