+++ CHANGELOG.md
@@ -4,9 +4,13 @@ Release Notes
 
 ## Next Version
 
+**New**
+
+- Better counting support in the query interface ([documentation](#fetching-aggregated-values))
+
 **Breaking Changes**
 
-- Swift enums that behave like other database values now need to declare `DatabaseValueConvertible` conformance. The `DatabaseIntRepresentable`, `DatabaseInt32Representable`, `DatabaseInt64Representable` and `DatabaseStringRepresentable` have been removed ([documentation](https://github.com/groue/GRDB.swift#swift-enums))
+- Swift enums that behave like other database values now need to declare `DatabaseValueConvertible` adoption. The `DatabaseIntRepresentable`, `DatabaseInt32Representable`, `DatabaseInt64Representable` and `DatabaseStringRepresentable` protocols have been removed ([documentation](https://github.com/groue/GRDB.swift#swift-enums))
 
 
 ## 0.45.1